[MH3-60104] production hex files are added for all the configurations and mc3 files are updated

+++ apps/ethercat_counter_foe_app/firmware/src/config/same53_lan9255_evb/peripheral/eic/plib_eic.c
@@ -137,12 +137,12 @@ void EIC_Initialize (void)
 
 void EIC_InterruptEnable (EIC_PIN pin)
 {
-    EIC_REGS->EIC_INTENSET = (1UL << (uint32_t)pin);
+    EIC_REGS->EIC_INTENSET = (1UL << pin);
 }
 
 void EIC_InterruptDisable (EIC_PIN pin)
 {
-    EIC_REGS->EIC_INTENCLR = (1UL << (uint32_t)pin);
+    EIC_REGS->EIC_INTENCLR = (1UL << pin);
 }
 
 void EIC_CallbackRegister(EIC_PIN pin, EIC_CALLBACK callback, uintptr_t context)
